ut2/ut2003/4->ut3: modified default scale factor, added UTTeamStaticMesh all: fixed overidded materials not being converted properly for staticmeshes fixed staticmesh uv refactor using component class for light conversion

+++ src/test/java/org/xtx/ut4converter/tools/psk/PSKStaticMeshTest.java
@@ -50,18 +50,32 @@ public class PSKStaticMeshTest {
 
     @Test
     void testReadAndExportToASE() throws URISyntaxException, IOException {
-        final File pskFile = new File(Objects.requireNonNull(PSKStaticMeshTest.class.getResource("/meshes/DirtChunk_01aw.pskx")).toURI());
 
-        File t3dFile = File.createTempFile("DirtChunk_01aw", "ase");
-        //File t3dFile = new File("C:\\dev3\\TEMP2\\DirtChunk_01aw.ase");
-        FileUtils.delete(t3dFile);
+        // uncomment to test conversion locally (no t3d file delete for testing import in editor)
+        final File pskFile = new File(Objects.requireNonNull(PSKStaticMeshTest.class.getResource("/meshes/Cube256.pskx")).toURI());
+        boolean isLocalTest = false;
+        File t3dFile;
+
+        if (isLocalTest) {
+            t3dFile = new File("C:\\Temp\\UT-Converter\\Cube256.ase");
+        } else {
+            t3dFile = File.createTempFile("Cube256", "ase");
+        }
 
         try {
             final PSKStaticMesh pskStaticMesh = new PSKStaticMesh(pskFile);
+            pskStaticMesh.getMaterials().get(0).setMaterialName("Tex256x256-2_Mat");
+            pskStaticMesh.getMaterials().get(1).setMaterialName("Tex256x256-3_Mat");
+            pskStaticMesh.getMaterials().get(2).setMaterialName("Tex256x256-4_Mat");
+            pskStaticMesh.getMaterials().get(3).setMaterialName("Tex256x256-1_Mat");
+            pskStaticMesh.getMaterials().get(4).setMaterialName("Tex256x256-5_Mat");
+            pskStaticMesh.getMaterials().get(5).setMaterialName("Tex256x256-6_Mat");
             pskStaticMesh.exportToAse(t3dFile);
             Assertions.assertTrue(t3dFile.length() > 0);
         } finally {
-            FileUtils.deleteQuietly(t3dFile);
+            if (!isLocalTest) {
+                FileUtils.deleteQuietly(t3dFile);
+            }
         }
     }
 }
